change field type of SizeRange from u32 -> u16

+++ src/size_range.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct SizeRange {
-    pub(crate) min: u32,
-    pub(crate) max: u32,
+    pub(crate) min: u16,
+    pub(crate) max: u16,
 }
